# postman/CarteiraBank.Services.Api.openapi.v1.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "CarteiraBank.Services.Api | v1",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:5109/"
    }
  ],
  "paths": {
    "/api/contratos/{id}/acordos": {
      "post": {
        "tags": [
          "Acordos"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CriarAcordoRequisicao"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CriarAcordoRequisicao"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CriarAcordoRequisicao"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/acordos/{id}/boleto": {
      "post": {
        "tags": [
          "Acordos"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/boletos/{id}/pdf": {
      "get": {
        "tags": [
          "Boletos"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/clientes/me": {
      "get": {
        "tags": [
          "Clientes"
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/contratos": {
      "get": {
        "tags": [
          "Contratos"
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/contratos/{id}/saldo-devedor": {
      "get": {
        "tags": [
          "Contratos"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/solicitacoes-credito": {
      "post": {
        "tags": [
          "Credito"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CriarSolicitacaoCreditoRequisicao"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CriarSolicitacaoCreditoRequisicao"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CriarSolicitacaoCreditoRequisicao"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "get": {
        "tags": [
          "Credito"
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/solicitacoes-credito/{id}/decisao": {
      "post": {
        "tags": [
          "Credito"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DecidirSolicitacaoCreditoRequisicao"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/DecidirSolicitacaoCreditoRequisicao"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/DecidirSolicitacaoCreditoRequisicao"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/health": {
      "get": {
        "tags": [
          "Health"
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "CriarAcordoRequisicao": {
        "required": [
          "valorDesconto",
          "quantidadeParcelas"
        ],
        "type": "object",
        "properties": {
          "valorDesconto": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?$",
            "type": [
              "number",
              "string"
            ],
            "format": "double"
          },
          "quantidadeParcelas": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          }
        }
      },
      "CriarSolicitacaoCreditoRequisicao": {
        "required": [
          "valorSolicitado",
          "quantidadeParcelasSolicitada",
          "finalidade"
        ],
        "type": "object",
        "properties": {
          "valorSolicitado": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?$",
            "type": [
              "number",
              "string"
            ],
            "format": "double"
          },
          "quantidadeParcelasSolicitada": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "finalidade": {
            "type": "string"
          }
        }
      },
      "DecidirSolicitacaoCreditoRequisicao": {
        "required": [
          "aprovar",
          "motivo"
        ],
        "type": "object",
        "properties": {
          "aprovar": {
            "type": "boolean"
          },
          "motivo": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      }
    }
  },
  "tags": [
    {
      "name": "Acordos"
    },
    {
      "name": "Boletos"
    },
    {
      "name": "Clientes"
    },
    {
      "name": "Contratos"
    },
    {
      "name": "Credito"
    },
    {
      "name": "Health"
    }
  ]
}